% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTesting.R
\name{hypothesisTesting}
\alias{hypothesisTesting}
\title{Extended part of Hypothesis testing}
\usage{
hypothesisTesting(data, vect = c(), y = NULL, mu = NULL,
  tail = "two.sided", alpha = 0.05)
}
\arguments{
\item{data=}{dataframe}

\item{vect=}{list of column number to be pass}

\item{y=}{target variable (Optional)}

\item{mu=}{Population mean if single popultion test need to be performed}

\item{tail=}{two.sided/greater/less(if not provided "two.sided" will be default)}

\item{aplha=}{provide significance value (default=0.05)}
}
\description{
Will do hypothesis testing as per user requirements
}
\author{
Tejas Vartak and Shivani Goyal
}
